+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 =================
 
 ## 0.18.70
+### is.codion.common.reactive
+- Value.Builder.changeConsumer() and weakChangeConsumer() wildcard removed from type parameter, was unusable.
 ### is.codion.swing.common.ui
 - DefaultFilterTableCellEditor.DefaultCellEditable now prevents editing from starting on key event if modifiers are enabled.
 ### is.codion.framework.domain